# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Majstorce API",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://api.majstorce.mk/api/latest",
            "description": "Production Enviroment"
        },
        {
            "url": "http://api.majstorce.mk.local/api/latest",
            "description": "Local Enviroment - with host"
        },
        {
            "url": "http://localhost/api/latest",
            "description": "Local Enviroment -  docker"
        }
    ],
    "paths": {
        "/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "List categories  filtered by tags",
                "description": "List categories  filtered by tags",
                "operationId": "get_categories",
                "parameters": [
                    {
                        "name": "find-by-tag",
                        "in": "query",
                        "description": "Find category by tag",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "array list of categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#components/schemas/ResponseCategory"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Create new category",
                "description": "Create new category",
                "operationId": "Categories::index_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/RequestCategory"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Insert CAtegory was saccesfull",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ResponseCategory"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ResponseFail"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/categories/{categoryId}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "List category by categoryId",
                "description": "List category by categoryId",
                "operationId": "get_category",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Category id to be listed",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ResponseCategory"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Category"
                ],
                "summary": "Update category by categoryID",
                "description": "Update category by categoryID",
                "operationId": "Categories::index_put",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Category id to update",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/RequestCategory"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Category update was accepted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ResponseCategory"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ResponseFail"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Delete a category",
                "description": "Delete Category By CategoryID",
                "operationId": "Categories::index_delete",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "Category id to delete",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "array list of categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#components/schemas/ResponseCategory"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/client/personal-details": {
            "get": {
                "tags": [
                    "Client"
                ],
                "description": "Cliant personal details",
                "operationId": "PersonalDetails::getPersonalDetails_get",
                "responses": {
                    "200": {
                        "description": "Registration was saccesfull",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ResponseClientPersonalDetails"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ResponseFail"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "Precondition failed - Invalid Data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ResponseFail"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/oauth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "description": "Cliant Login",
                "operationId": "Login::login_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/RequestClientLogin"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registration was saccesfull",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ResponseAuthorizationTokens"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid username or passsword",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ResponseFail"
                                }
                            }
                        }
                    },
                    "412": {
                        "description": "Precondition failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/ResponseFail"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorizati3on",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}